// src/app/types/LiveReels.ts
export interface LiveList {
    status:  boolean;
    data:    LiveData;
    message: string;
    toast:   boolean;
}

export interface LiveData {
    Records:    LiveRecord[];
    Pagination: LivePagination;
}

export interface LivePagination {
    total_pages:      number;
    total_records:    number;
    current_page:     number;
    records_per_page: number;
}

export interface LiveRecord {
    live_id:        number;
    total_viewers:  number;
    curent_viewers: number;
    likes:          number;
    socket_room_id: string;
    comments:       number;
    live_title:     string;
    live_status:    string;
    is_demo:        boolean;
    createdAt:      string;
    updatedAt:      string;
    Live_hosts:     LiveHost[];
}

export interface LiveHost {
    live_host_id: number;
    is_live:      boolean;
    is_main_host: boolean;
    peer_id:      string;
    createdAt:    string;
    updatedAt:    string;
    live_id:      number;
    user_id:      number;
    User:         LiveUser;
    following:    boolean;
}

export interface LiveUser {
    user_name:                   string;
    email:                       string;
    mobile_num:                  string;
    profile_pic:                 string;
    platforms:                   string[];
    user_id:                     number;
    full_name:                   string;
    first_name:                  string;
    last_name:                   string;
    country_code:                string;
    login_type:                  string;
    country:                     string;
    country_short_name:          string;
    profile_verification_status: boolean;
    total_socials:               number;
    blocked_by_admin:            boolean;
    is_deleted:                  boolean;
    createdAt:                   string;
    updatedAt:                   string;
}